truthfirst-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
pillow==10.1.0
weasyprint==59.0
jinja2==3.1.2
matplotlib==3.10.8
aiosmtplib==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
requests==2.32.5
transformers==4.57.3
pytesseract==0.3.10
email-validator==2.3.0
reportlab==4.1.0
pydyf==0.7.0